requirements.txt
bitarray==2.5.1
paho-mqtt==1.6.1
tabulate==0.8.9
